ai/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
redis==5.0.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# AI/ML Libraries
langchain==0.1.0
langchain-openai==0.0.2
langchain-google-genai==0.0.5
langchain-anthropic==0.0.1
openai>=1.6.1,<2.0.0
google-generativeai==0.3.2
anthropic>=0.8.0,<0.9.0

# Additional AI libraries
chromadb==0.4.18
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.1.1
numpy==1.24.3
pandas==2.1.4

# Utilities
python-dateutil==2.8.2
aiofiles==23.2.1
tenacity==8.2.3
